ALSA: constify snd_pcm_ops structures
Check for snd_pcm_ops structures that are only stored in the ops field
of a snd_soc_platform_driver structure or passed as the third argument
to snd_pcm_set_ops.  The corresponding field or parameter is declared
const, so snd_pcm_ops structures that have this property can be declared
as const also.

This is a backport from mainline commit
6769e988b0062bf8d63d35b0fdbb47b385dd3252